+++ myapp/simple-console-app/src/main/java/com/learn/simpleconsoleapp/seedworks/advices/SecurityAdvice.java
@@ -1,4 +1,4 @@
-package com.learn.simpleconsoleapp.seedworks;
+package com.learn.simpleconsoleapp.seedworks.advices;
 
 import com.learn.simpleconsoleapp.services.SecurityManager;
 import org.springframework.aop.MethodBeforeAdvice;
